refactor(validation): replace TaskLogStream with ValidationLog component
