+++ apps/console/src/features/tasks/components/TaskPage.tsx
@@ -40,8 +40,7 @@ import { TaskLogView } from "./TaskLogView";
 // can take a minute) reads as "still working" rather than a stall. The clock
 // restarts whenever a new line arrives or the stream (re)connects, and stops
 // ticking (no wasted 1s re-renders) once there is nothing to wait on.
-// Exported for ValidationPage, which shares the same log anatomy.
-export function useSecondsSince(resetKey: string, active: boolean): number {
+function useSecondsSince(resetKey: string, active: boolean): number {
   const [seconds, setSeconds] = useState(0);
   useEffect(() => {
     setSeconds(0);
@@ -56,7 +55,7 @@ export function useSecondsSince(resetKey: string, active: boolean): number {
   return seconds;
 }
 
-export const EXEC_ACTIVE = new Set(["queued", "running"]);
+const EXEC_ACTIVE = new Set(["queued", "running"]);
 
 // The per-task console (#173): slim header (title, status chip, GitHub
 // icon-link) over the flat streaming log. get-task provides the initial
